+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- (notification) detailed pomodoro status in system notification [#204](https://github.com/sectore/timr-tui/pull/204)
 - (cli) Add `--auto-switch` argument [#203](https://github.com/sectore/timr-tui/pull/203)
 
 ## v1.10.0 - 2026-06-16